+++ src/onegov.activity/onegov/activity/matching/__init__.py
@@ -2,11 +2,12 @@ from onegov.activity.matching.core import deferred_acceptance
 from onegov.activity.matching.core import deferred_acceptance_from_database
 from onegov.activity.matching.interfaces import MatchableBooking
 from onegov.activity.matching.interfaces import MatchableOccasion
-from onegov.activity.matching.score import Scoring
-from onegov.activity.matching.score import PreferMotivated
-from onegov.activity.matching.score import PreferInAgeBracket
-from onegov.activity.matching.score import PreferOrganiserChildren
 from onegov.activity.matching.score import PreferAdminChildren
+from onegov.activity.matching.score import PreferGroups
+from onegov.activity.matching.score import PreferInAgeBracket
+from onegov.activity.matching.score import PreferMotivated
+from onegov.activity.matching.score import PreferOrganiserChildren
+from onegov.activity.matching.score import Scoring
 
 __all__ = [
     'deferred_acceptance',
@@ -14,6 +15,7 @@ __all__ = [
     'MatchableBooking',
     'MatchableOccasion',
     'Scoring',
+    'PreferGroups',
     'PreferMotivated',
     'PreferInAgeBracket',
     'PreferOrganiserChildren',